// src/pages/adminMenu/inputUserRef.interface.ts
export interface IInputUserRef{
    inputEmail: React.RefObject<HTMLInputElement>
    inputPassword: React.RefObject<HTMLInputElement>
    inputSurname: React.RefObject<HTMLInputElement>
    inputName: React.RefObject<HTMLInputElement>
    roleSelect: React.RefObject<HTMLSelectElement>
    inputAccess: React.RefObject<HTMLInputElement>,
}